@@ -11,13 +11,10 @@ abstract class FilesystemTask extends BaseTask
 {
     use LoadFilesystemTasks;
 
-    /**
-     * {@inheritdoc}
-     */
     public function run()
     {
         $task = static::NAME;
-        $arguments = \array_values($this->getTaskArguments());
+        $arguments = array_values($this->getTaskArguments());
 
         return $this->task(Filesystem::class)->{$task}(...$arguments)->run();
     }
