@@ -7,9 +7,8 @@ namespace PhpTaskman\CoreTasks\Plugin\Task;
 use PhpTaskman\CoreTasks\Plugin\BaseTask;
 use Robo\Exception\TaskException;
 
-/**
- * Class BaseConfiguration.
- */
+use function is_array;
+
 abstract class BasePhpTask extends BaseTask
 {
     /**
@@ -48,12 +47,12 @@ abstract class BasePhpTask extends BaseTask
 
         $line['stop'] = $blockEnd;
 
-        $textChecksum = \sha1(\json_encode($line));
+        $textChecksum = sha1(json_encode($line));
 
         $line['start'] .= '(' . $textChecksum . ')';
         $line['stop'] .= '(' . $textChecksum . ')';
 
-        return \implode("\n", $line);
+        return implode("\n", $line);
     }
 
     /**
@@ -71,16 +70,16 @@ abstract class BasePhpTask extends BaseTask
      */
     protected function getStatement($variable, $name, $value)
     {
-        $output = \var_export($value, true);
+        $output = var_export($value, true);
 
-        if (\is_array($value)) {
-            $output = \str_replace(
+        if (is_array($value)) {
+            $output = str_replace(
                 [' ', "\n", '=>', ',)', '),'],
                 ['', '', ' => ', ')', '), '],
                 $output
             );
         }
 
-        return \sprintf("$%s['%s'] = %s;", $variable, $name, $output);
+        return sprintf("$%s['%s'] = %s;", $variable, $name, $output);
     }
 }
